tschnoor - all three assignments (draft) (#32)
## Summary of Changes

Here is my initial draft post for all three assignments.

**Paper Summary (draft)**: Added a citation for my selected paper. Also
added a brief description of the paper and wrote about my motivation for
choosing it.

**Course Project (draft)**: Provided a description of the goals of my
proposed project, as well as a timeline for when I hope to get things
done and a description of how I will be limiting the scope appropriately
for this class. I will be working alone on this.

**Class Competition (draft)**: Added my Kaggle info and initial
leaderboard results. Also described my current features and
architecture, as well as my general modeling approach going forward. I
will be working alone on this one too.

_I just want to point out that this repository includes files created by
a student with user ID **sydneyb**. I think that maybe they pushed to
main before I created my branch. It isn't a problem for me, I just
wanted to make it clear that I didn't create that repository or the
files in it._