Initial commit for gquinones (#22)
## Summary of Changes

Greetings! Here I am uploading all of the new changes:

- For the summary of the article, I didn't change anything.
- For the competition, I: changed the description, made it brief and
concise; changed the title so it better reflects what the file is about;
and moved what was previously in the description to the approach
section.
- For the class project, I: uploaded the information to include my
GitHub name; since I'm working with Jake Mains, I added him as an
author; I created a better title to better reflect the information on
the file; included the link to the GitHub repo with the code and
approach; and, in the index.mdx file, added the approach and timeline of
this project.

Please let me know if something should be fixed!